Conflicts:
	src/stan/agrad/fwd/matrix/dot_product.hpp
	src/stan/agrad/matrix.hpp
	src/stan/gm/function_signatures.h
	src/test/agrad/fwd/abs_test.cpp
	src/test/agrad/fwd/fabs_test.cpp
	src/test/agrad/fwd/log_sum_exp_test.cpp
	src/test/agrad/fwd/matrix/sort_test.cpp
	src/test/gm/model_specs/compiled/function_signatures1.stan

+++ src/stan/mcmc/hmc/hamiltonians/dense_e_point.hpp
@@ -27,9 +27,9 @@ namespace stan {
       void write_metric(std::ostream* o) {
         if(!o) return;
         *o << "# Elements of inverse mass matrix:" << std::endl;
-        for(size_t i = 0; i < mInv.rows(); ++i) {
+        for(int i = 0; i < mInv.rows(); ++i) {
           *o << "# " << mInv(i, 0) << std::flush;
-          for(size_t j = 1; j < mInv.cols(); ++j)
+          for(int j = 1; j < mInv.cols(); ++j)
             *o << ", " << mInv(i, j) << std::flush;
           *o << std::endl;
         }
